Write smart contract
This commit adds the Adoption smart contract to the project.
Adoption.sol includes two functions:

adopt() - Will facilitate users making adoption requests. We use the require() statement to ensure the ID is within range. If the ID is in range, we then add the address that made the call to our adopters array and return the petId provided as a confirmation.

getAdopters() - Returns adopters Array.This will allow our UI to have access to all pet adoption statuses instead of calling each one individually.